styleAI-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
scikit-learn>=1.3.0,<1.5.0
torch==2.0.0
numpy==1.24.4
requests==2.31.0
matplotlib==3.8.0
transformers==4.31.0